requirements.txt
chardet
requests
urllib3
python-nmap
argparse
socket
IPy
re